rmbg-api/requirements.txt
fastapi
uvicorn[standard]
rembg==2.0.56
Pillow
numpy
